EnhancedPasswordManager/requirements.txt
bcrypt
pycryptodome
pandas
openpyxl
pytest
tkinter
flask